Gross hack to fix Atom urls
Pages don't have a URL set until they're routed to a file, obviously.
However, the whole point of compiling each post twice is to use
a simpler template the second time, aggregate the posts in the "feed"
group, and write them to a single atom.xml file instead of individual
files on disk. Thus, the individual pages don't have a $url$ property.
This breaks the Atom template, leaving literal "$url$"s all over. Solved
by a nasty hack - compile the "feed" group and the default group with
the same route, overwriting the feed-optimized pages with the nicely
styled ones.